Try to return creature's raw XP value if level is not available for PF2E difficulty calc
Using pf2e difficulty calculations doesn't work when using inline creatures that specify their XP value. There is no (documented) way to specify level in the inline encounter builder. Since the difficulty logic attempts to use the lvl field to re-derive the xp, an xp/difficulty of "0" is reported for inline-defined creatures.

This change adds an intermediate fallback of returning `creature.xp` if the lvl is unavailable, but still falls back to `0` if `xp` is also undefined/null